Major v2 migration: restructure app, migrate API calls, overhaul resource detail pages
- Reorganise code into routes/, utils/, views/ packages; remove flat-file layout
- Migrate all client calls to nano-crucible v2 resource API (datasets, samples,
  projects, instruments, graphs); drop all deprecated flat-method calls
- Add count() usage for dashboard stats (avoids full list fetches)
- Fix _paginate to support limit=None for fetching all records
- Fix include_links=True on samples.list() and normalise datasets from links field
- Fix KeyError on optional fields (measurement, sample_type) with .get()
- Add scientific metadata section to sample detail page with tree search
- Fetch scientific metadata per-sample on detail page (include_metadata=True)
- Split ancestors/descendants into direct parents/children vs further relatives
- Extract shared Jinja2 macros (mrow, mrow_link, section_metadata, deletion_banner)
- Move toggleDetailsAdv() to base.html; Bootstrap Icons CDN consolidated in base
- Resource detail pages: clickable Project and Owner ORCID fields, deletion warning
  banner, data_type field, field layout aligned with mobile app reference design
- Fix global search thread-safety (pass client explicitly to avoid current_app proxy)
