+++ flask_templates/entity_graph.html
@@ -44,9 +44,9 @@
         <i class="bi bi-image me-1"></i>Hide Thumbnails
     </button>
     <span class="text-muted small">
-        <span class="legend-dot" style="background:#4a7ba7;"></span>Sample
+        <span class="legend-dot" style="background:var(--cy-sample-color);"></span>Sample
         &nbsp;
-        <span class="legend-dot" style="background:#5a9e6f;"></span>Dataset
+        <span class="legend-dot" style="background:var(--cy-dataset-color);"></span>Dataset
     </span>
 </div>
