+++ flask_templates/instrument_list.html
@@ -10,14 +10,8 @@
 
 {% block content %}
 <style>
-    .list-row {
-        display: flex; align-items: center; gap: 0.75rem;
-        padding: 0.6rem 0.75rem;
-        border-bottom: 1px solid var(--bs-border-color);
-        transition: background 0.1s; text-decoration: none; color: inherit;
-    }
-    .list-row:last-child { border-bottom: none; }
-    .list-row:hover { background: var(--bs-tertiary-bg); color: inherit; }
+    /* icon + text layout: tighter horizontal padding, wider gap than default .list-row */
+    .list-row { gap: 0.75rem; padding: 0.55rem 0.75rem; }
 </style>
 
 <div style="max-width:720px; margin: 0 auto;">
