+++ flask_templates/mdnote_edit.html
@@ -12,7 +12,7 @@
         }
         #mdnote-editor {
             flex: 1;
-            border: 1px solid #ddd;
+            border: 1px solid var(--bs-border-color);
             border-radius: 4px;
             overflow: hidden;
         }
@@ -23,9 +23,9 @@
             flex: 1;
             overflow-y: auto;
             padding: 1rem 1.5rem;
-            border: 1px solid #ddd;
+            border: 1px solid var(--bs-border-color);
             border-radius: 4px;
-            background: white;
+            background: var(--bs-body-bg);
             line-height: 1.6;
         }
         #mdnote-preview h1, #mdnote-preview h2, #mdnote-preview h3,
@@ -35,7 +35,7 @@
             font-weight: 600;
         }
         #mdnote-preview h1, #mdnote-preview h2 {
-            border-bottom: 1px solid #eaecef;
+            border-bottom: 1px solid var(--bs-border-color);
             padding-bottom: 0.3em;
         }
         #mdnote-preview p { margin-bottom: 1em; }
@@ -50,7 +50,7 @@
             font-size: 85%;
         }
         #mdnote-preview pre {
-            background: #f6f8fa;
+            background: var(--bs-tertiary-bg);
             border-radius: 6px;
             padding: 1em;
             overflow: auto;
@@ -58,10 +58,10 @@
         }
         #mdnote-preview pre code { background: transparent; padding: 0; }
         #mdnote-preview blockquote {
-            border-left: 4px solid #dfe2e5;
+            border-left: 4px solid var(--bs-border-color);
             padding-left: 1em;
             margin-left: 0;
-            color: #6a737d;
+            color: var(--bs-secondary-color);
         }
         #mdnote-preview table {
             border-collapse: collapse;
@@ -69,12 +69,12 @@
             margin-bottom: 1em;
         }
         #mdnote-preview th, #mdnote-preview td {
-            border: 1px solid #dfe2e5;
+            border: 1px solid var(--bs-border-color);
             padding: 6px 13px;
         }
-        #mdnote-preview th { background: #f6f8fa; font-weight: 600; }
-        .wiki-link-sample { color: #2d6a4f; font-weight: 500; }
-        .wiki-link-dataset { color: #1d3557; font-weight: 500; }
+        #mdnote-preview th { background: var(--bs-tertiary-bg); font-weight: 600; }
+        .wiki-link-sample { color: var(--bs-success); font-weight: 500; }
+        .wiki-link-dataset { color: var(--bs-primary); font-weight: 500; }
     </style>
 {% endblock %}
