+++ flask_templates/users.html
@@ -24,17 +24,11 @@
         border-left-width: 4px;
         border-radius: 0.375rem 0.375rem 0 0;
     }
-    .list-row {
-        display: flex; align-items: center; gap: 0.75rem;
-        padding: 0.55rem 0.75rem;
-        border-bottom: 1px solid var(--bs-border-color);
-        transition: background 0.1s; text-decoration: none; color: inherit;
-    }
-    .list-row:last-child { border-bottom: none; }
-    .list-row:hover { background: var(--bs-tertiary-bg); color: inherit; }
+    /* avatar + text layout: wider gap, tighter horizontal padding than default .list-row */
+    .list-row { gap: 0.75rem; padding: 0.55rem 0.75rem; }
 </style>
 
-<div style="max-width: 680px; margin: 0 auto;">
+<div style="max-width: 720px; margin: 0 auto;">
 
 <div class="d-flex align-items-center justify-content-between mb-3 flex-wrap gap-2">
     <h1 class="fs-4 fw-semibold mb-0">
@@ -84,18 +78,19 @@
         {% set first    = m.get('first_name') or '' %}
         {% set last     = m.get('last_name')  or '' %}
         {% set name     = (first ~ ' ' ~ last).strip() %}
-        {% set email    = m.get('lbl_email') or m.get('email') or '' %}
+        {% set email    = m.get('email') or '' %}
         {% set initials = ((first[:1] if first else '') ~ (last[:1] if last else '')) | upper or '?' %}
-        {% if m.get('orcid') %}
-        <a class="list-row" href="/user/{{ m['orcid'] }}"
-           data-search="{{ (name ~ ' ' ~ email ~ ' ' ~ m['orcid']).lower() }}"
-           data-orcid="{{ m['orcid'] }}">
+        {% set uid = m.get('unique_id') %}
+        {% if uid %}
+        <a class="list-row" href="/user/{{ uid }}"
+           data-search="{{ (name ~ ' ' ~ email ~ ' ' ~ uid).lower() }}"
+           data-orcid="{{ uid }}">
         {% else %}
         <div class="list-row"
              data-search="{{ (name ~ ' ' ~ email).lower() }}"
              data-orcid="">
         {% endif %}
-            <div class="user-avatar" data-avatar-for="{{ name or m.get('orcid') or '?' }}">
+            <div class="user-avatar" data-avatar-for="{{ name or uid or '?' }}">
                 {{ initials }}
             </div>
             <div style="min-width:0; flex:1;">
@@ -104,10 +99,10 @@
                 <div class="text-muted small">{{ email }}</div>
                 {% endif %}
             </div>
-            {% if m.get('orcid') %}
-            <span class="mfid small text-nowrap">{{ m['orcid'] }}</span>
+            {% if uid %}
+            <span class="mfid small text-nowrap">{{ uid }}</span>
             {% endif %}
-        {% if m.get('orcid') %}</a>{% else %}</div>{% endif %}
+        {% if uid %}</a>{% else %}</div>{% endif %}
         {% endfor %}
     {% else %}
         <div class="px-3 py-2 text-muted small">No member information available.</div>
@@ -118,15 +113,6 @@
 </div>
 
 <script>
-    const PALETTE = [
-        '#4e79a7','#f28e2b','#e15759','#76b7b2','#59a14f',
-        '#b07aa1','#9c755f','#6b6ecf','#e05c5c','#3a9e7e'
-    ];
-    function hashColor(str) {
-        let h = 0;
-        for (let i = 0; i < str.length; i++) h = (Math.imul(31, h) + str.charCodeAt(i)) | 0;
-        return PALETTE[Math.abs(h) % PALETTE.length];
-    }
     document.querySelectorAll('[data-hash-color]').forEach(el => {
         el.style.borderLeftColor = hashColor(el.dataset.hashColor);
     });
