+++ routes/instruments.py
@@ -1,0 +1,38 @@
+import logging
+
+import flask
+from flask import Blueprint, abort, render_template
+
+logger = logging.getLogger(__name__)
+
+
+def create_blueprint(auth):
+    bp = Blueprint('instruments_routes', __name__)
+
+    @bp.route("/instruments/")
+    @auth.oidc_auth('orcid')
+    def instrument_list():
+        instruments = flask.current_app.crucible_client.instruments.list()
+        return render_template('instrument_list.html', instruments=instruments)
+
+    @bp.route("/instrument/<instrument_id>")
+    @auth.oidc_auth('orcid')
+    def instrument_detail(instrument_id):
+        import views.instruments as instrument_views
+        client = flask.current_app.crucible_client
+        instrument = client.instruments.get(instrument_id=instrument_id)
+        if not instrument:
+            abort(404)
+        instrument_name = instrument.get('instrument_name', '')
+        custom_views = instrument_views.get_views(instrument_name, instrument_id)
+        recent_datasets = []
+        if instrument_name:
+            try:
+                recent_datasets = client.datasets.list(instrument_name=instrument_name, limit=None)
+                recent_datasets.sort(key=lambda d: d.get('timestamp') or '', reverse=True)
+            except Exception:
+                recent_datasets = []
+        return render_template('instrument.html', instrument=instrument,
+                               custom_views=custom_views, recent_datasets=recent_datasets)
+
+    return bp
