+++ views/datasets/pollux_oospec.py
@@ -122,10 +122,10 @@ def _parse_h5_position(content_bytes, pos_name):
 
 def _fetch_h5_bytes(dsid):
     """Fetch the first .h5 associated file for the dataset and return its bytes."""
-    ds = current_app.crucible_client.get_dataset(dsid)
-    associated_files = current_app.crucible_client.get_associated_files(dsid)
+    ds = current_app.crucible_client.datasets.get(dsid)
+    associated_files = current_app.crucible_client.datasets.get_associated_files(dsid)
     try:
-        download_links = current_app.crucible_client.get_dataset_download_links(dsid)
+        download_links = current_app.crucible_client.datasets.get_download_links(dsid)
     except Exception:
         download_links = {}
 
@@ -158,7 +158,7 @@ def create_blueprint(auth, helpers):
     def view(project_id, dsid):
         if not is_user_in_project(project_id):
             abort(403)
-        ds = current_app.crucible_client.get_dataset(dsid)
+        ds = current_app.crucible_client.datasets.get(dsid)
         return render_template('dataset_views/pollux_oospec.html',
                                project_id=project_id, ds=ds)
