+++ views/instruments/ncem_team.py
@@ -15,13 +15,13 @@ def create_blueprint(auth, helpers):
     @bp.route('/overview')
     @auth.oidc_auth('orcid')
     def overview():
-        instrument = current_app.crucible_client.get_instrument(instrument_name='team05')
+        instrument = current_app.crucible_client.instruments.get(instrument_name='team05')
         if not instrument:
            abort(404)
         #instrument_name = instrument.get('instrument_name', '')
 
         # TODO see if user has access to this instrument
-        datasets = current_app.crucible_client.list_datasets(
+        datasets = current_app.crucible_client.datasets.list(
             instrument_name='team05', limit=500
         )
         datasets.sort(key=lambda d: d.get('timestamp') or '', reverse=True)
