+++ views/instruments/nirvana.py
@@ -11,11 +11,11 @@ def create_blueprint(auth, helpers):
     @bp.route('/overview')
     @auth.oidc_auth('orcid')
     def view():
-        instrument = current_app.crucible_client.get_instrument(instrument_id='nirvana')
+        instrument = current_app.crucible_client.instruments.get(instrument_id='nirvana')
         #if not instrument:
         #    abort(404)
         #instrument_name = instrument.get('instrument_name', '')
-        datasets = current_app.crucible_client.list_datasets(
+        datasets = current_app.crucible_client.datasets.list(
             instrument_name='nirvana spectrometer', limit=500
         )
         datasets.sort(key=lambda d: d.get('timestamp') or '', reverse=True)
